Fix compiler warnings and add a warnings-as-errors CI gate
Clean up the warnings surfaced by the MSVC and Clang builds and add a
lint gate so they cannot regress.

Warning fixes:
- Fix four real 64-bit bugs flagged as MSVC C4334: `1 << n` is a 32-bit
  shift but the result is stored in a 64-bit usize_t with n up to 63
  (xdelta3-hash.h, xdelta3-second.h). Use `(usize_t)1 << n`.
- Add explicit casts at the ~38 intentional narrowing conversions MSVC
  reports as C4244/C4267 (codec internals writing bytes/indices), across
  xdelta3.c and the djw/fgk/hash/blkcache/main/test headers. No behavior
  change; the truncations are now documented.
- Replace the incompatible `(xd3_posix_func*) &read/&write` casts with
  thin wrapper functions, fixing -Wcast-function-type on GCC and Clang.
- Define _CRT_SECURE_NO_WARNINGS on MSVC to silence the C4996 "unsafe"
  nags about standard functions like snprintf.

Lint gate:
- Add an XD3_WERROR option that turns on /WX (MSVC) and -Werror (Clang)
  for the core targets (xdelta3, xdelta3decode, xdelta3link). GCC is
  excluded: it emits many additional, largely false-positive diagnostics
  (implicit-fallthrough, format-truncation, stringop-overflow) that are
  out of scope. The C++ test harness and vendored cpp-btree are also
  excluded as they carry their own pre-existing warnings.
- Enable XD3_WERROR=ON in the CMake and Windows CI jobs.

Also bump actions/checkout from v4 to v5.

+++ xdelta3/linkxd3lib.c
@@ -35,8 +35,8 @@ int main() {
   use (xd3_encode_stream (& stream, NULL, 0, NULL, NULL, 0));
   xd3_set_appheader (& stream, NULL, 0);
   use (xd3_encoder_used_source (& stream));
-  use (xd3_encoder_srcbase (& stream));
-  use (xd3_encoder_srclen (& stream));
+  use ((int) xd3_encoder_srcbase (& stream));
+  use ((int) xd3_encoder_srclen (& stream));
 #endif
   return 0;
 }
